feat: 添加简单的fork系统调用
添加了fork系统调用，当前功能还不完善，且有些以前commit留下的bug
添加了getpid系统调用，获取当前进程的pid
添加了一个用于调试的系统调用debug
更改pid_t的定义，将其类型改为int

+++ include/alphaz/type.h
@@ -10,7 +10,7 @@ typedef unsigned long   u64;
 
 typedef unsigned long   size_t;
 
-typedef unsigned short  pid_t;
+typedef int             pid_t;
 
 typedef long            ssize_t;
